Extract Class - Refactoring
