+++ core/src/main/java/pro/zackpollard/telegrambot/api/TelegramBot.java
@@ -41,7 +41,7 @@ import java.util.Map;
 /**
  * @author Zack Pollard
  */
-public final class TelegramBot {
+public final class TelegramBot extends TelegramCallBot{
 
     /**
      * The API URL endpoint that is constant for all bots
@@ -59,7 +59,7 @@ public final class TelegramBot {
     private final Map<Class<?>, Extension.Provider<? extends Extension>> extensions = new HashMap<>();
 
     @Getter
-    private final String authToken;
+    public static String authToken = null;
     private final ListenerRegistry listenerRegistry;
     private UpdateManager updateManager = null;
 
@@ -154,7 +154,7 @@ public final class TelegramBot {
      *
      * @return The URL used to make all the API calls to the bot API
      */
-    public String getBotAPIUrl() {
+    public static String getBotAPIUrl() {
 
         return API_URL + "bot" + authToken + "/";
     }
@@ -460,7 +460,7 @@ public final class TelegramBot {
                 break;
             }
             case VIDEO_NOTE: {
-                
+
                 SendableVideoNoteMessage videoNoteMessage = (SendableVideoNoteMessage) message;
 
                 try {
@@ -896,11 +896,11 @@ public final class TelegramBot {
 
         return false;
     }
-    
+
     public boolean deleteMessage(Message message) {
         return this.deleteMessage(message.getChat().getId(), message.getMessageId());
     }
-    
+
     public boolean deleteMessage(String chatId, Long messageId) {
         if(chatId != null && messageId != null) {
 
@@ -927,104 +927,6 @@ public final class TelegramBot {
         return false;
     }
 
-    /**
-     * This allows you to respond to an inline query with an InlineQueryResponse object
-     *
-     * @param inlineQueryId         The ID of the inline query you are responding to
-     * @param inlineQueryResponse   The InlineQueryResponse object that you want to send to the user
-     *
-     * @return True if the response was sent successfully, otherwise False
-     */
-    public boolean answerInlineQuery(String inlineQueryId, InlineQueryResponse inlineQueryResponse) {
-
-        if (inlineQueryId != null && inlineQueryResponse != null) {
-
-            HttpResponse<String> response;
-            JSONObject jsonResponse;
-
-            try {
-                MultipartBody requests = Unirest.post(getBotAPIUrl() + "answerInlineQuery")
-                        .field("inline_query_id", inlineQueryId)
-                        .field("results", GSON.toJson(inlineQueryResponse.getResults()))
-                        .field("cache_time", inlineQueryResponse.getCacheTime())
-                        .field("is_personal", inlineQueryResponse.isPersonal())
-                        .field("next_offset", inlineQueryResponse.getNextOffset())
-                        .field("switch_pm_text", inlineQueryResponse.getSwitchPmText())
-                        .field("switch_pm_parameter", inlineQueryResponse.getSwitchPmParameter());
-
-                response = requests.asString();
-                jsonResponse = Utils.processResponse(response);
-
-                if (jsonResponse != null) {
-
-                    if (jsonResponse.getBoolean("result")) return true;
-                }
-            } catch (UnirestException e) {
-                e.printStackTrace();
-            }
-        }
-
-        return false;
-    }
-
-    /**
-     * This allows you to respond to a callback query with some text as a response. This will either show up as an
-     * alert or as a toast on the telegram client
-     *
-     * @param callbackQueryId   The ID of the callback query you are responding to
-     * @param text              The text you would like to respond with
-     * @param showAlert         True will show the text as an alert, false will show it as a toast notification
-     *
-     * @deprecated  This method is deprecated in favour of the {@link #answerCallbackQuery(String, CallbackQueryResponse)}
-     *              method, this should be used for all new implementations
-     *
-     * @return True if the response was sent successfully, otherwise False
-     */
-    @Deprecated
-    public boolean answerCallbackQuery(String callbackQueryId, String text, boolean showAlert) {
-
-        return this.answerCallbackQuery(callbackQueryId, CallbackQueryResponse.builder().text(text).showAlert(showAlert).build());
-    }
-
-    /**
-     * This allows you to respond to a callback query with some text as a response. This will either show up as an
-     * alert or as a toast on the telegram client
-     *
-     * @param callbackQueryId       The ID of the callback query you are responding to
-     * @param callbackQueryResponse The response that you would like to send in reply to this callback query
-     *
-     * @return True if the response was sent successfully, otherwise False
-     */
-    public boolean answerCallbackQuery(String callbackQueryId, CallbackQueryResponse callbackQueryResponse) {
-
-        if(callbackQueryId != null && callbackQueryResponse.getText() != null) {
-
-            HttpResponse<String> response;
-            JSONObject jsonResponse;
-
-            try {
-                MultipartBody requests = Unirest.post(getBotAPIUrl() + "answerCallbackQuery")
-                        .field("callback_query_id", callbackQueryId, "application/json; charset=utf8;")
-                        .field("text", callbackQueryResponse.getText(), "application/json; charset=utf8;")
-                        .field("show_alert", callbackQueryResponse.isShowAlert())
-                        .field("cache_time", callbackQueryResponse.getCacheTime())
-                        .field("url", callbackQueryResponse.getURL() != null ? callbackQueryResponse.getURL().toExternalForm() : null, "application/json; charset=utf8;");
-
-
-                response = requests.asString();
-                jsonResponse = Utils.processResponse(response);
-
-                if (jsonResponse != null) {
-
-                    if (jsonResponse.getBoolean("result")) return true;
-                }
-            } catch (UnirestException e) {
-                e.printStackTrace();
-            }
-        }
-
-        return false;
-    }
 
     /**
      * Use this method to kick a user from a group or a supergroup. In the case of supergroups, the user will not be
@@ -1197,3 +1099,4 @@ public final class TelegramBot {
         return listenerRegistry;
     }
 }
+
